changed TreeCode/forceTree.cpp -- changed kappa in force2D from double to float
 changed TreeCode_link/Tree.cpp
-- introduced a pointer check after a realloc call

 changed TreeCode_link/curve_routines.cpp
-- added a function for priting the caustic and critical curves

+++ TreeCode/forceTree.cpp
@@ -208,7 +208,7 @@ float * ForceTree::CalculateSPHsmoothing(int N){
  *       4*pi*G*mass_scale to get the deflection angle caused by the plane lens.
  * */
 
-void ForceTree::force2D(double *ray,double *alpha,double *kappa,double *gamma,bool no_kappa){
+void ForceTree::force2D(double *ray,double *alpha,float *kappa,float *gamma,bool no_kappa){
 
   PosType xcm,ycm,rcm2,tmp;
   int OpenBox(TreeNBHndl tree,PosType r);
